Refactor SQL integration
