+++ homeassistant/components/sql/sensor.py
@@ -7,19 +7,11 @@ import decimal
 import logging
 from typing import Any
 
-import sqlalchemy
-from sqlalchemy import lambda_stmt
 from sqlalchemy.engine import Result
 from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.orm import Session, scoped_session, sessionmaker
-from sqlalchemy.sql.lambdas import StatementLambdaElement
-from sqlalchemy.util import LRUCache
+from sqlalchemy.orm import scoped_session
 
-from homeassistant.components.recorder import (
-    CONF_DB_URL,
-    SupportedDialect,
-    get_instance,
-)
+from homeassistant.components.recorder import CONF_DB_URL, get_instance
 from homeassistant.components.sensor import CONF_STATE_CLASS
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import (
@@ -29,12 +21,10 @@ from homeassistant.const import (
     CONF_UNIQUE_ID,
     CONF_UNIT_OF_MEASUREMENT,
     CONF_VALUE_TEMPLATE,
-    EVENT_HOMEASSISTANT_STOP,
     MATCH_ALL,
 )
-from homeassistant.core import Event, HomeAssistant, callback
+from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import TemplateError
-from homeassistant.helpers import issue_registry as ir
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import (
     AddConfigEntryEntitiesCallback,
@@ -50,12 +40,15 @@ from homeassistant.helpers.trigger_template_entity import (
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
 from .const import CONF_ADVANCED_OPTIONS, CONF_COLUMN_NAME, CONF_QUERY, DOMAIN
-from .models import SQLData
-from .util import redact_credentials, resolve_db_url
+from .util import (
+    async_create_sessionmaker,
+    generate_lambda_stmt,
+    redact_credentials,
+    resolve_db_url,
+    validate_query,
+)
 
 _LOGGER = logging.getLogger(__name__)
-
-_SQL_LAMBDA_CACHE: LRUCache = LRUCache(1000)
 
 TRIGGER_ENTITY_OPTIONS = (
     CONF_AVAILABILITY,
@@ -145,36 +138,6 @@ async def async_setup_entry(
     )
 
 
-@callback
-def _async_get_or_init_domain_data(hass: HomeAssistant) -> SQLData:
-    """Get or initialize domain data."""
-    if DOMAIN in hass.data:
-        sql_data: SQLData = hass.data[DOMAIN]
-        return sql_data
-
-    session_makers_by_db_url: dict[str, scoped_session] = {}
-
-    #
-    # Ensure we dispose of all engines at shutdown
-    # to avoid unclean disconnects
-    #
-    # Shutdown all sessions in the executor since they will
-    # do blocking I/O
-    #
-    def _shutdown_db_engines(event: Event) -> None:
-        """Shutdown all database engines."""
-        for sessmaker in session_makers_by_db_url.values():
-            sessmaker.connection().engine.dispose()
-
-    cancel_shutdown = hass.bus.async_listen_once(
-        EVENT_HOMEASSISTANT_STOP, _shutdown_db_engines
-    )
-
-    sql_data = SQLData(cancel_shutdown, session_makers_by_db_url)
-    hass.data[DOMAIN] = sql_data
-    return sql_data
-
-
 async def async_setup_sensor(
     hass: HomeAssistant,
     trigger_entity_config: ConfigType,
@@ -187,70 +150,16 @@ async def async_setup_sensor(
     async_add_entities: AddEntitiesCallback | AddConfigEntryEntitiesCallback,
 ) -> None:
     """Set up the SQL sensor."""
-    try:
-        instance = get_instance(hass)
-    except KeyError:  # No recorder loaded
-        uses_recorder_db = False
-    else:
-        uses_recorder_db = db_url == instance.db_url
-    sessmaker: scoped_session | None
-    sql_data = _async_get_or_init_domain_data(hass)
-    use_database_executor = False
-    if uses_recorder_db and instance.dialect_name == SupportedDialect.SQLITE:
-        use_database_executor = True
-        assert instance.engine is not None
-        sessmaker = scoped_session(sessionmaker(bind=instance.engine, future=True))
-    # For other databases we need to create a new engine since
-    # we want the connection to use the default timezone and these
-    # database engines will use QueuePool as its only sqlite that
-    # needs our custom pool. If there is already a session maker
-    # for this db_url we can use that so we do not create a new engine
-    # for every sensor.
-    elif db_url in sql_data.session_makers_by_db_url:
-        sessmaker = sql_data.session_makers_by_db_url[db_url]
-    elif sessmaker := await hass.async_add_executor_job(
-        _validate_and_get_session_maker_for_db_url, db_url
-    ):
-        sql_data.session_makers_by_db_url[db_url] = sessmaker
-    else:
+    (
+        sessmaker,
+        uses_recorder_db,
+        use_database_executor,
+    ) = await async_create_sessionmaker(hass, db_url)
+    if sessmaker is None:
         return
+    validate_query(hass, query_str, uses_recorder_db, unique_id)
 
     upper_query = query_str.upper()
-    if uses_recorder_db:
-        redacted_query = redact_credentials(query_str)
-
-        issue_key = unique_id if unique_id else redacted_query
-        # If the query has a unique id and they fix it we can dismiss the issue
-        # but if it doesn't have a unique id they have to ignore it instead
-
-        if (
-            "ENTITY_ID," in upper_query or "ENTITY_ID " in upper_query
-        ) and "STATES_META" not in upper_query:
-            _LOGGER.error(
-                "The query `%s` contains the keyword `entity_id` but does not "
-                "reference the `states_meta` table. This will cause a full table "
-                "scan and database instability. Please check the documentation and use "
-                "`states_meta.entity_id` instead",
-                redacted_query,
-            )
-
-            ir.async_create_issue(
-                hass,
-                DOMAIN,
-                f"entity_id_query_does_full_table_scan_{issue_key}",
-                translation_key="entity_id_query_does_full_table_scan",
-                translation_placeholders={"query": redacted_query},
-                is_fixable=False,
-                severity=ir.IssueSeverity.ERROR,
-            )
-            raise ValueError(
-                "Query contains entity_id but does not reference states_meta"
-            )
-
-        ir.async_delete_issue(
-            hass, DOMAIN, f"entity_id_query_does_full_table_scan_{issue_key}"
-        )
-
     # MSSQL uses TOP and not LIMIT
     if not ("LIMIT" in upper_query or "SELECT TOP" in upper_query):
         if "mssql" in db_url:
@@ -271,39 +180,6 @@ async def async_setup_sensor(
             )
         ],
     )
-
-
-def _validate_and_get_session_maker_for_db_url(db_url: str) -> scoped_session | None:
-    """Validate the db_url and return a session maker.
-
-    This does I/O and should be run in the executor.
-    """
-    sess: Session | None = None
-    try:
-        engine = sqlalchemy.create_engine(db_url, future=True)
-        sessmaker = scoped_session(sessionmaker(bind=engine, future=True))
-        # Run a dummy query just to test the db_url
-        sess = sessmaker()
-        sess.execute(sqlalchemy.text("SELECT 1;"))
-
-    except SQLAlchemyError as err:
-        _LOGGER.error(
-            "Couldn't connect using %s DB_URL: %s",
-            redact_credentials(db_url),
-            redact_credentials(str(err)),
-        )
-        return None
-    else:
-        return sessmaker
-    finally:
-        if sess:
-            sess.close()
-
-
-def _generate_lambda_stmt(query: str) -> StatementLambdaElement:
-    """Generate the lambda statement."""
-    text = sqlalchemy.text(query)
-    return lambda_stmt(lambda: text, lambda_cache=_SQL_LAMBDA_CACHE)
 
 
 class SQLSensor(ManualTriggerSensorEntity):
@@ -329,7 +205,7 @@ class SQLSensor(ManualTriggerSensorEntity):
         self.sessionmaker = sessmaker
         self._attr_extra_state_attributes = {}
         self._use_database_executor = use_database_executor
-        self._lambda_stmt = _generate_lambda_stmt(query)
+        self._lambda_stmt = generate_lambda_stmt(query)
         if not yaml and (unique_id := trigger_entity_config.get(CONF_UNIQUE_ID)):
             self._attr_name = None
             self._attr_has_entity_name = True
